Hwb color space
Add HWB color space, basically following the spec. It's been spot-tested with non-HDR colors.

Serialization is to syntax which can be parsed by CSS Color 4, but doesn't include `%` for W and B, as is done by color.js. It also doesn't strictly follow serialization by the spec, which calls for the color to be lowered to RGB. To me this is an argument in favor of having a separate code path for strict CSS serialization.

In theory, HWB should be able to handle HDR colors with negative values for W and B. In practice, that's a bad idea, and in any case the current code does not roundtrip these correctly. That might be a flaw in the spec, and deserves deeper investigation.

Closes #19