Add Adobe RGB (1998) color space
The Linear A98 RGB <-> Linear sRGB matrices are based on the exact
rational conversion matrices calculated from the XYZ chromaticities.
This crate doesn't need the exact rationals as they're truncated to f32
at compile-time anyway, but it might be worth having to serve as a
reference for other people. The same could be done for the other basis
change matrices.

(All based on the assumption that the color spaces' specified xy
chromaticities are normative, and the specified matrices in the specs
are computed off those.)