Pack color name into `DynamicColor` flags
A rough implementation of the ideas in
https://github.com/linebender/color/issues/26#issuecomment-2467997421.

The ergonomics require some tweaking, and perhaps the performance too,
though it should be easy to write this in a form where the compiler can
easily optimize the bitwise operations. I haven't verified what the
compiler does with the current form.

An alternative is to keep `Missing` unchanged, and add something similar
to `Flags` as a new field on `DynamicColor`. But adding a new field
makes constructing `DynamicColor` a bit less ergonomic.

---------

Co-authored-by: Daniel McNab <36049421+DJMcNab@users.noreply.github.com>