Add conversion roundtrip tests
This tests roundtripping equality of all the rectangular color spaces.
The precision bound is relatively tight: currently arbitrarily chosen as
16 float epsilons. This allows doing many tests without manually
specifying the expected values.

I've added a test for Oklab as well, but all color spaces can be added
in time. (I'm out of time for today.) We probably do need to think a
little bit about the test cases that are interesting to hit. This now
tests somewhat arbitrary values, but it would be worthwhile to hit
component values near the piece edges of transfer functions, or near the
hue edges, etc.