Optionally support `serde`.
This adds support for `serde::Deserialize` and `serde::Serialize` using
`serde-derive` for `AlphaColor`, `OpaqueColor`, `PremulColor`, `Rgba8`.

This does not provide it for `DynamicColor` to avoid dealing with
the non-exhaustive `ColorSpaceTag`.

It uses `derive` rather than hand-implementing the traits as the
latter is very verbose and most things using `serde` will already
have enabled the `derive` feature. (This is also true for the
other Linebender crates.)