Rely on defined saturation behavior for f32 to u8 casts
This is a backport of a change from Raph made in the CPU
sparse strips experiment.

When converting from `f32` to `u8`, the code was clamping to
be sure that the values were within range. This was necessary
in the past, but hasn't been needed for some years now, since
<https://github.com/rust-lang/rust/issues/10184>. This clamping
is not necessary and the compiler is not optimizing it away
as of Rust 1.83 on Apple Silicon.