Always return little-endian packed value in Rgba8::to_u32
If the intention is for the bytes of r, g, b and a to always be in the
same memory order, I believe this patch is required.

In the order of memory of `to_u8_array`, the red channel is at the
lowest memory address and therefore (in little endian) the least
significant byte. On a big endian system, using `from_le_bytes` would
move the red byte to the highest memory addres.