First draft of Okhsv color space
This is a mostly mechanical translation from the
[reference implementation provided by Björn Ottosson](https://github.com/bottosson/bottosson.github.io/blob/f6f08b7fde9436be1f20f66cebbc739d660898fd/misc/ok_color.h).

I've not done a deep dive into the math. This just differs by a few
optimizations.

Note: the conversion into and out of this color space are approximations
due to `compute_max_srgb_saturation`. The other picker, `Okhsl`, would
share most of this code.