Reuse XYZ-D65's conversions for XYZ-D50 absolute conversions
More specifically: for the absolute conversions to/from linear sRGB.

This works because XYZ-D65 and sRGB both use the D65 white point, so the
"relative" conversion specified in XYZ-D65 to/from linear sRGB is an
absolute one. Plus, though XYZ-D50 is relative to D50, XYZ-D50 and
XYZ-D65 are otherwise the same color space.